backend/requirements.txt
# Cognivista QA Backend - Dependencies
# =====================================

# Core Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Database
motor>=3.3.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# File Upload
python-multipart>=0.0.9

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# AWS Bedrock
boto3>=1.34.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Data Processing (optional - for future features)
numpy>=1.24.0
pandas>=2.0.0

# Audio Processing (for transcription)
openai-whisper>=20231117
pydub>=0.25.1
ffmpeg-python>=0.2.0
